Refactor to simplify C/R services A whole bunch of other improvements were made. For more details check https://linear.app/cedana/issue/CED-317/simplify-cedana-interface-layer-cliapi

+++ api/runc/runc.go
@@ -33,7 +33,7 @@ func List(root string) error {
 	return nil
 }
 
-func GetPidByContainerId(containerId, root string) (int, error) {
+func GetPidByContainerId(containerId, root string) (int32, error) {
 	statePath := filepath.Join(root, containerId, "state.json")
 	var runcSpec libcontainer.State
 	if _, err := os.Stat(statePath); err == nil {
@@ -45,7 +45,7 @@ func GetPidByContainerId(containerId, root string) (int, error) {
 			return 0, err
 		}
 	}
-	return runcSpec.InitProcessPid, nil
+	return int32(runcSpec.InitProcessPid), nil
 }
 
 func GetContainerIdByName(containerName string, root string) (string, string, error) {
